Only parse when we have translations and dont clear cache before post

+++ api/src/Service/ObjectEntityService.php
@@ -540,7 +540,9 @@ class ObjectEntityService
         }
 
         // If we change an ObjectEntity we should remove it from the result cache
-        $this->functionService->removeResultFromCache($objectEntity);
+        if ($this->request->getMethod() != 'POST') {
+            $this->functionService->removeResultFromCache($objectEntity);
+        }
         return $objectEntity;
     }
